moved tests into individual unit test files
