+++ tests/units/API/instanceRegion.js
@@ -1,0 +1,40 @@
+module.exports = function(expect, eventEmitter, validCivo, invalidCivo, civoStub) {
+  describe('listRegions()', () => {
+    it('valid auth', (done) => {
+      Promise.all([
+        eventEmitter.once('received'),
+        validCivo.listRegions()
+      ]).then((data) => {
+        const request = data[0][0];
+        const response = data[1];
+        expect(request.status).to.be.equal(200, 'returned status should be 200');
+        expect(request.method).to.be.equal('GET', 'listRegions() should be a GET request');
+        expect(request.url).to.be.equal('/regions', 'listRegions() should call "/regions" endpoint');
+        expect(Object.keys(request.body)).to.have.lengthOf(0, 'No body data should be recived');
+        expect(Object.keys(request.params)).to.have.lengthOf(0, 'No params should be used');
+        expect(JSON.stringify(response)).to.be.equal(JSON.stringify(civoStub.responses.getRegions.response), 'correct response was not returned');
+        done();
+      }).catch((err) => {
+        done(err);
+      });
+    });
+    it('invalid auth', (done) => {
+      Promise.all([
+        eventEmitter.once('received'),
+        invalidCivo.listRegions()
+      ]).then((data) => {
+        const request = data[0][0];
+        const response = data[1];
+        expect(request.status).to.be.equal(401, 'returned status should be 401 unauthorised');
+        expect(request.method).to.be.equal('GET', 'listRegions() should be a GET request');
+        expect(request.url).to.be.equal('/regions', 'listRegions() should call "/regions" endpoint');
+        expect(Object.keys(request.body)).to.have.lengthOf(0, 'No body data should be recived');
+        expect(Object.keys(request.params)).to.have.lengthOf(0, 'No params should be used');
+        expect(JSON.stringify(response)).to.be.equal(JSON.stringify(civoStub.errors.authentication), 'correct response was not returned');
+        done();
+      }).catch((err) => {
+        done(err);
+      });
+    });
+  });
+};
