+++ tests/units/API/instances.js
@@ -1,0 +1,67 @@
+module.exports = function(expect, eventEmitter, validCivo, invalidCivo, civoStub) {
+  describe('listInstances()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('createInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('deleteInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('getInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('retagInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('rebootInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('shutdownInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('startInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('resizeInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('rebuiltInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('restoreInstance()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('updateInstanceFirewall()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+  describe('moveInstancePublicIp()', () => {
+    it('force fail', () => {
+      expect(false).to.be.true;
+    });
+  });
+};
